Make "ancient" Gnus look like a regular "read" message
This is in response to the issue 119 by sivaramn: <https://github.com/protesilaos/modus-themes/issues/119>.
In the original Gnus faces, the "ancient" are not the same as "read"
in terms of their colour, though I cannot find what the difference
actually is. It seems to me that "ancient" simply means "read it before".