Reverse theme rotation direction when called with a prefix
When calling modus-themes-rotate interactively and with a prefix
argument, reverse the sense of rotation from right to left among the
modus-themes-to-rotate.  Explicit calls to modus-themes-rotate can
pass a non-nil REVERSE argument to achieve the same for their THEMES.

* modus-themes.el: (modus-themes-rotate, modus-themes--rotate-p,
modus-themes--rotate, modus-themes--next-in-rotation):

Add an optional REVERSE argument to modus-themes-rotate.  When called
interactively this becomes the current prefix argument.  Replace
helper functions with a single modus-themes--next-in-rotation.  Update
doc strings.