Bump version number to trigger GNU ELPA package update
I do not want to publish a new minor release because I did that
recently and there are not many changes since then. Plus, I would need
to prepare a change log, which I cannot do right now.

Rebuilding the package allows us to include commit bf6c6ce, which
lowers the minimum required version of Emacs to 27.1. Hopefully,
everything will work as intended.

This is in response to bug#78991 by John Sullivan:
<https://lists.gnu.org/archive/html/bug-gnu-emacs/2025-07/msg00480.html>.