Revert "Add an 'eval-and-compile' around all palette variables"
This reverts commit 7111708f15256cdb107e5441cdd2fb4cc709c7e5.

I thought this was needed to fix a warning I get about 'nil'
foreground value for a couple of faces. It seems that the mappings in
'modus-themes-common-palette-mappings' are not read and so I suspected
this was because of them not being known at compile time. Apparently,
this does not fix the problem, so I must check elsewhere...