BREAKING Be much more conservative with :inherit
This is related to Emacs bug#79672: <https://debbugs.gnu.org/cgi/bugreport.cgi?bug=79672>.

What I did was to review ALL the uses of :inherit and try to do the
following:

- Limit recursive inheritence.
- Inherit from Modus faces.
- Do not inherit when semantic palette mappings can be used directly.

There are some exceptions, but the idea is to rely less on :inherit
and more on the palette entries we provide. This is anyway good for
us: I needed to make this change since I introduce the semantic
palette mappings.

In the process of doing this I also reviewed lots of faces. I might
have made some mistakes, but things are looking good overall.